// src/shared/api/generated/schemas/article.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * DOCUMENTATION
 * OpenAPI spec version: 1.0.0
 */
import type { ArticleAuthor } from './articleAuthor';
import type { ArticleCategory } from './articleCategory';
import type { ArticleCover } from './articleCover';
import type { ArticleCreatedBy } from './articleCreatedBy';
import type { ArticleLocalizationsItem } from './articleLocalizationsItem';
import type { ArticleUpdatedBy } from './articleUpdatedBy';

export interface Article {
  articleContent?: unknown;
  author?: ArticleAuthor;
  category?: ArticleCategory;
  cover?: ArticleCover;
  createdAt?: string;
  createdBy?: ArticleCreatedBy;
  description?: string;
  documentId?: string;
  id?: number;
  locale?: string;
  localizations?: ArticleLocalizationsItem[];
  publishedAt?: string;
  slug?: string;
  title?: string;
  updatedAt?: string;
  updatedBy?: ArticleUpdatedBy;
}